+++ tests/back/Common/EntityWithValue/Code.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Test\Common\Builder\EntityWithValue;
+namespace Akeneo\Test\Common\EntityWithValue;
 
 final class Code
 {
